+++ CHANGELOG.md
@@ -1,7 +1,9 @@
-0.0.2
------
-- Removed autoloader and debug stuff
+0.0.2 (unreleased)
+------------------
+- Added replacing parameters in translation strings
 - Added key prefix
+- Removed autoloader and debug stuff
+- Removed default parameter
 
 0.0.1
 -----